Rename "admin" folder to "admin_views"
Same with "base", once there are any files in it
