+++ editorsnotes_app/src/admin_views/document/index.js
