+++ editorsnotes_app/src/admin_views/document/scan_list.js
