+++ editorsnotes_app/src/admin_views/document/templates/document.html
