+++ editorsnotes_app/src/admin_views/document/templates/zotero_item.html
