+++ editorsnotes_app/src/admin_views/document/templates/zotero_item_type_select.html
