+++ editorsnotes_app/src/admin_views/document/zotero.js
