+++ editorsnotes_app/src/admin_views/generic/make_modal_view.js
