+++ editorsnotes_app/src/admin_views/generic/ordered_collection_base.js
