+++ editorsnotes_app/src/admin_views/generic/select_item_base.js
