+++ editorsnotes_app/src/admin_views/generic/templates/add_item_modal.html
