+++ editorsnotes_app/src/admin_views/generic/templates/add_or_select_item.html
