+++ editorsnotes_app/src/admin_views/generic/templates/confirm_delete.html
