+++ editorsnotes_app/src/admin_views/note/index.js
