+++ editorsnotes_app/src/admin_views/note/note_section.js
