+++ editorsnotes_app/src/admin_views/note/templates/note_section.js
