+++ editorsnotes_app/src/admin_views/topic/templates/citation_list.html
