+++ editorsnotes_app/src/admin_views/widgets/feedback/modal.js
