+++ editorsnotes_app/src/admin_views/widgets/select_note.js
